feat(auth): extract auth types to separate file
BREAKING CHANGE: AuthResult and StaffUser types moved from staff.ts to types.ts
