+++ app/actions/auth/types.ts
@@ -1,0 +1,17 @@
+// Define consistent return type
+export interface AuthResult<T = unknown> {
+  data: T | null;
+  error: {
+    code: string;
+    message: string;
+  } | null;
+}
+
+export interface StaffUser {
+  id: string;
+  email: string;
+  role?: string;
+  collection: string;
+  createdAt: string;
+  updatedAt: string;
+}
